requirements.txt
kaggle==1.6.17
pyyaml==6.0.2
python-dotenv==1.0.1
pydantic==2.9.2
tenacity==8.2.3
requests==2.31.0
python-dateutil==2.9.0
flask==3.0.0
